Former-commit-id: 7829fdd2ee2ebe1fc67f0d5cc80ed1103a62bcdb

+++ scripts/sdsc_localize_pids.py
@@ -1,24 +1,36 @@
-# %%
-import warnings
-warnings.simplefilter("ignore", category=DeprecationWarning)
-
-import numpy as np
-from one.api import ONE
-import spikeinterface.core as sc
-import spikeinterface.extractors as se
-from spike_psvae import subtract
-from pathlib import Path
+import argparse
 import pickle
 import shutil
-import argparse
 import subprocess
+import warnings
+from pathlib import Path
+
+import h5py
+import numpy as np
+import one.alf.io as alfio
+import spikeinterface.core as sc
+import spikeinterface.extractors as se
 from brainbox.io.one import SpikeSortingLoader
-from spike_psvae.spike_train_utils import clean_align_and_get_templates
+from dredge import dredge_ap
+from dredge import motion_util as mu
+from one.api import ONE
+from spike_psvae import subtract
 from spike_psvae.grab_and_localize import grab_and_localize
-import spikeglx
+from spike_psvae.spike_train_utils import clean_align_and_get_templates
+
+warnings.simplefilter("ignore", category=DeprecationWarning)
 
 
 sdsc_base_path = Path("/mnt/sdceph/users/ibl/data")
+
+
+def save_motion_est(path, name, me):
+    with open(path / f"{name}_motion_est.pkl", "wb") as jar:
+        pickle.dump(me, jar)
+
+
+def rsync(src, dest):
+    subprocess.run(["rsync", "-avP", str(src), str(dest)])
 
 
 def eid2sdscpath(eid):
@@ -33,7 +45,9 @@ def eid2sdscpath(eid):
         assert len(rel_path) == 1
         rel_path = Path(rel_path[0])
         searchdir = (
-            sdsc_base_path / alyx_base_path.relative_to(one.cache_dir) / rel_path.parent
+            sdsc_base_path
+            / alyx_base_path.relative_to(one.cache_dir)
+            / rel_path.parent
         )
         pattern = Path(rel_path.name).with_suffix(f".*.cbin")
         glob = list(searchdir.glob(str(pattern)))
@@ -53,7 +67,11 @@ if __name__ == "__main__":
     ap.add_argument("--batchlen", type=float, default=1)
     ap.add_argument("--locworkers", type=int, default=2)
     ap.add_argument("--ksreloc", action="store_true")
+    ap.add_argument("--ksmotion", action="store_true")
+    ap.add_argument("--residnorm-decrease", type=float, default=np.sqrt(10.0))
     ap.add_argument("--tmp-parent", type=str, default="/tmp")
+    ap.add_argument("--no-overwrite", action="store_true")
+    ap.add_argument("--rerun-errors", action="store_true")
 
     args = ap.parse_args()
 
@@ -109,9 +127,15 @@ if __name__ == "__main__":
                             metadata,
                         )
                     # continue
-                if "subtraction_error" in meta:
-                    print("This one had a problem during subtraction in a previous run. Skipping")
+                if "subtraction_error" in meta and not args.rerun_errors:
+                    print(
+                        "This one had a problem during subtraction in a previous run. Skipping"
+                    )
                     continue
+                if "subtraction_error" in meta and args.rerun_errors:
+                    print(
+                        "This one had a problem during subtraction in a previous run. Rerunning."
+                    )
 
         with open(sessdir / "metadata.pkl", "wb") as sess_jar:
             pickle.dump(metadata, sess_jar)
@@ -157,7 +181,7 @@ if __name__ == "__main__":
         destriped_bin = dscache / f"destriped_{cbin_rel}"
         destriped_bin = destriped_bin.with_suffix(".bin")
         if not destriped_bin.exists():
-            shutil.copyfile(cbin_path, dscache / cbin_rel)
+            rsync(cbin_path, dscache / cbin_rel)
             shutil.copyfile(meta_path, dscache / meta_rel)
             shutil.copyfile(ch_path, dscache / ch_rel)
 
@@ -167,6 +191,7 @@ if __name__ == "__main__":
                         "python",
                         str(Path(__file__).parent / "destripe.py"),
                         str(dscache / cbin_rel),
+                        f"--n-jobs={args.njobs}",
                     ],
                     check=True,
                 )
@@ -184,7 +209,7 @@ if __name__ == "__main__":
                 #         (dscache / pfile).unlink()
 
         assert destriped_bin.exists()
-        
+
         rec = sc.read_binary(
             destriped_bin,
             rec_cbin.get_sampling_frequency(),
@@ -199,6 +224,7 @@ if __name__ == "__main__":
             end_frame=rec.get_num_samples() // 2 + 1000,
         )
         print(f"{ttt.min()=} {ttt.max()=}")
+        
 
         # if subcache.exists():
         #     shutil.rmtree(subcache)
@@ -206,6 +232,9 @@ if __name__ == "__main__":
         if not (sessdir / "subtraction.h5").exists():
             print("Subtracting")
             try:
+                if np.abs(ttt).max() > 1e5:
+                    print("Data values are weird!")
+                    raise ValueError("skip")
                 sub_h5 = subtract.subtraction(
                     rec,
                     out_folder=subcache,
@@ -219,8 +248,65 @@ if __name__ == "__main__":
                     overwrite=False,
                     n_sec_chunk=args.batchlen,
                     save_cleaned_pca_projs_on_n_channels=5,
-                    loc_feature=("ptp", "peak"),
+                    loc_feature=("ptpgpu"),
+                    residnorm_decrease=args.residnorm_decrease,
                 )
+                with h5py.File(sub_h5, "r+") as h5:
+                    h5.create_dataset(
+                        "localizations", data=h5["localizationsptpgpu"][:]
+                    )
+                    spike_times = h5["spike_index"][:, 0] + (
+                        h5["start_time"][()] * 30_000
+                    )
+                    x = h5["localizations"][:, 0]
+                    z = h5["localizations"][:, 2]
+                    maxptp = h5["maxptps"][:]
+                    geom = h5["geom"][:]
+                    t = spike_times / 30000
+
+                    wh = (
+                        (z > geom[:, 1].min() - 250)
+                        & (z < geom[:, 1].max() + 250)
+                        & (x > geom[:, 0].min() - 250)
+                        & (x < geom[:, 0].max() + 250)
+                    )
+
+                    tme, _ = dredge_ap.register(
+                        maxptp[wh],
+                        z[wh],
+                        t[wh],
+                        max_disp_um=50,
+                        pbar=False,
+                    )
+                    save_motion_est(sessdir, "dredge_ap", tme)
+                    z_reg = tme.correct_s(spike_times / 30000, z)
+                    h5.create_dataset("z_reg", data=z_reg)
+
+                    # ks' drift correction
+                    if args.ksmotion:
+                        ssl = SpikeSortingLoader(one=one, pid=pid)
+                        ssl.download_spike_sorting_object("drift")
+                        drift = alfio.load_object(
+                            ssl.files["drift"], wildcards=ssl.one.wildcards
+                        )
+                        drift_samples = ssl.samples2times(drift["times"], direction="reverse")
+
+                        # code from pyks. get the centers of the bins that they used
+                        nblocks = (drift["um"].shape[1] + 1) // 2
+                        yl = np.floor(geom[:, 1].max() / nblocks).astype("int") - 1
+                        mins = np.linspace(0, geom[:, 1].max() - yl - 1, 2 * nblocks - 1)
+                        maxs = mins + yl
+                        centers = (mins + maxs) / 2
+                        print(f"{centers.shape=} {drift_samples.shape=} {drift['um'].shape=}")
+                        ksme = mu.NonrigidMotionEstimate(
+                            -drift["um"].T,
+                            time_bin_centers_s=drift_samples / fs,
+                            spatial_bin_centers_um=centers,
+                        )
+                        save_motion_est(sessdir, "ks", ksme)
+                        z_reg_ks = ksme.correct_s(t, z)
+                        h5.create_dataset("z_reg_ks", data=z_reg_ks)
+
                 shutil.copy(sub_h5, sessdir)
             except Exception as e:
                 with open(sessdir / "metadata.pkl", "wb") as sess_jar:
@@ -229,7 +315,8 @@ if __name__ == "__main__":
                 shutil.rmtree(dscache)
                 continue
             finally:
-                shutil.rmtree(subcache)
+                if subcache.exists():
+                    shutil.rmtree(subcache)
 
         if args.ksreloc and not (sessdir / "ks_relocalizations.npz").exists():
             print("Relocalizing KS")
@@ -237,9 +324,9 @@ if __name__ == "__main__":
             spikes, clusters, channels = sl.load_spike_sorting()
             clusters = sl.merge_clusters(spikes, clusters, channels)
             spike_times = spikes["times"]
-            spike_frames = sl.samples2times(spike_times, direction="reverse").astype(
-                "int"
-            )
+            spike_frames = sl.samples2times(
+                spike_times, direction="reverse"
+            ).astype("int")
             spike_train = np.c_[spike_frames, spikes["clusters"]]
 
             # -- align kilosort spikes
